alteraçoes 22.05

+++ pages/cursos/index.js
@@ -1,38 +1,37 @@
 import Pagina from "@/components/Pagina";
 import Link from "next/link";
 import React, { useEffect, useState } from "react";
-import { Table } from "react-bootstrap";
+import { Button, Table } from "react-bootstrap";
 import { AiFillPlusCircle } from "react-icons/ai";
 import { BiTrash } from "react-icons/bi";
+import { BsPencilFill } from "react-icons/bs";
 
 const index = () => {
+  const [cursos, setCursos] = useState([]);
 
-
-  const [cursos, setCursos] = useState([])
-  
   useEffect(() => {
-    setCursos(getAll())
-  }, [])
+    setCursos(getAll());
+  }, []);
 
-    function getAll(){
-      return JSON.parse(window.localStorage.getItem('cursos')) || []
+  function getAll() {
+    return JSON.parse(window.localStorage.getItem("cursos")) || [];
+  }
+
+  function excluir(id) {
+    if (confirm("Deseja realmente excluir o registro?")) {
+      const itens = getAll();
+      itens.splice(id, 1);
+      window.localStorage.setItem("cursos", JSON.stringify(itens));
+      setCursos(itens);
+      console.log(itens);
     }
-
-    function excluir(id){
-        const itens = getAll()
-        itens.splice(id, 1)
-        window.localStorage.setItem('cursos', JSON.stringify(itens))
-        setCursos(itens)
-        console.log(itens)
-      }
-
-
+  }
 
   return (
     <>
       <Pagina titulo="Cursos">
         <Link href="/cursos/form" className="btn btn-success">
-          <AiFillPlusCircle/> Novo
+          <AiFillPlusCircle /> Novo
         </Link>
         <br></br>
         <br></br>
@@ -48,12 +47,16 @@ const index = () => {
           <tbody>
             {cursos.map((item, i) => (
               <tr key={i}>
-              <td><BiTrash onClick={() => excluir(i)}/></td>
-              <td>{item.nome}</td>
-              <td>{item.duracao}</td>
-              <td>{item.modalidade}</td>
+                <td>
+                  <Link href={'/cursos/' + i}> </Link>
+                  <Button href={'/cursos/' + i}><BsPencilFill title="Alterar"/> </Button> 
+                  <Button><BiTrash onClick={() => excluir(i)}/> </Button>
+                </td>
+                <td>{item.nome}</td>
+                <td>{item.duracao}</td>
+                <td>{item.modalidade}</td>
               </tr>
-              ))}
+            ))}
           </tbody>
         </Table>
       </Pagina>
